Adding simple text to show min/max values for histogram

+++ Plugins/Render/GEK/Render/Device.hpp
@@ -539,7 +539,7 @@ namespace Gek
             size_t getHash(void) const;
         };
 
-        template <typename TYPE, uint8_t UNIQUE>
+        template <typename TYPE, size_t UNIQUE>
         struct Handle
         {
             TYPE identifier;
